Add ADC inputs.
Miraculously seems to "just work", even without any timing constraints,
even at 200 MHz! (using ADC's ramp pattern)
Need to beef up FIFO logic to be able to record long traces and check
for 100% accuracy, but it looks good.
Some issues seen with resetting the FIFOs from the control register - it
works in simulation, but not on the FPGA. However resetting the FIFOs
from the re-arm event seems to work fine.