Add CLKGEN programmability.
Not a full replacement for the S6 DCM yet, just direct R/W access to the
MMCM configuration registers. Seems to work well, as long as you don't
use fractional values, which are very badly documented (see XAPP888 and
the reference DRP design if you wish to go down that road).
For now, changed the default CLKGEN MMCM to *not* use fractional values.