Testbench: verify first sample value.
Previously the first sample value was assumed good, and only subsequent
values were verified. Now we peek at the ADC value when the trigger is
issued, so we can verify that the first received value is correct
(within some acceptable tolerance). Works with presamples too, *if* the
FIFO is out of reset when the trigger occurs.