FIFO management: ensure empty when done.
Easier said than done due to the 3-sample wide slow FIFO. Solution is to
capture extra samples to that the total number of captured samples is a
multiple of 3 -- otherwise last 1 or 2 samples could end up "trapped" in
the fast FIFO. Furthermore, we actually make the total number of
captured samples be a multiple of 6 (not 3), otherwise it's messy on the
slow FIFO read side (due to its 8-bit width). All this is under the hood
and everything should work out as long as the host sets the capture (or
segment) size to be a multiple of 6, and reads all that was captured.