Fix clip error logic (again).
Because of clock domain issues, it could fire if the logical OR of
successive samples was all 1's. Should be good and proper now. Also made
the clip error sticky, so that you'll know it happened even if it didn't
occur on your last capture (previously it was cleared by the next
capture).