SAD experiment: maximum single sample difference.
Tested new feature where in addition to the SAD threshold, there is a
"maximum deviation" parameter: if any single sample exceeds this
threshold, a match event is prevented (that SAD counter is marked
"dirty" until the offending sample has rolled out of the SAD counter).
I wanted to see whether this could reduce the jitter observed across
captures, but it doesn't seem to help. Kept the new code but it's
disabled by a configuration parameter, in case it's useful later. Also
note that it hasn't been fully tested.