Fix(?) FPGA read/write reliability issue.
During target debugging work, started seeing bitfiles which failed basic
register read/writes. The issue *seemed* to be with writing, since
reading fpga_buildtime was always fine. But haven't been able to catch
the problem live with ILAs (none of the bitfiles with ILAs exhibited the
problem). Suspect the issue is WRT relative timing of the CEN and WRN
signals, hence this fix, which *appears* to resolve the issue.