Detect post-DDR FIFO quasi-underflow.
If a FIFO read results in an empty FIFO (*not* at the very end of the
read), wrong data will be received since we're using the FIFO in
first-word-fall-through. So this is now caught and flagged as an
underflow. Probably quite unlikely to happen outside of simulation, but
even so it's handy to save the debugging time.