FIFO errors: (hopefully) fix (potential) CDC issue
Last build was starting to cause occasional presample errors in
test_husky.py's frequency sweep test; most of the FIFO errors originate
in the ADC clock domain but were being registered in the USB clock
domain without concern for CDC. This changes the registering to the ADC
clock domain, and appears to make the problem go away. Hopefully there
are no unintended consequences.