Address https://github.com/newaetech/chipwhisperer-husky-fpga/issues/17
Changed usb_slow_fifo's write clock from 200 MHz to 96 MHz; seems to
address the issue.

Note that changing the FIFO type from built-in to BRAM may be the
"proper" way to fix this, but leads to difficulties in timing closure
(especially for the Plus, which needs 250 MHz). So we rely on extensive
on-target testing to ensure this addresses the problem.