Fix for overflow when streaming.
One of the earlier commits (on the pro branch) broke this; the mechanism
of the bug (specifically, why it worked before) isn't fully understood,
but the symptom is that when a FIFO overflow occurs during a streaming
capture, the SAM3U gets a bit messed up because the capture doesn't
conclude as it should. The fix is that when a FIFO over/underflow error
occurs during a stream capture, stream_segment_available is held high
(regardless of FIFO contents), to ensure that SAM3U can finish reading.