Detect "trigger too soon" error in hardware.
This could happen if previous capture is not fully read; when the scope
is next armed, the unread capture needs to be flushed, which is now done
by reading the FIFO until it's empty (previously it was a FIFO reset but
Xilinx FIFOs are picky about resets, hence the change). If there are a
lot of unread samples, and the trigger comes very quickly after arming,
then the system might not be ready to capture and the trigger would be
missed. This error should flag that.