SAD: fix ref writing!
Finally works on-target, and implements nicely with the expected number
of ref samples. The trick was to eliminate indexing into refsamples for
writing, by using a FIFO. Makes writing a bit more fragile, and can no
longer be read, but performs beautifully and allow for a nice increase
in number of ref samples. May need some tweaks but appears to work quite
well on-target.