SAD fix for trigger sequencer.
Change the notion of "active" and add the notion of "trigger_allowed",
to deal with the SAD module's relatively long startup time which could
hurt usability in practice. Now, the SAD is kept running if it's part of
a trigger sequence, and triggers are simply blocked if they don't occur
during the window (marked by trigger_allowed).